% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrecipes-package.R
\docType{package}
\name{ggrecipes-package}
\alias{ggrecipes}
\alias{ggrecipes-package}
\title{ggrecipes: Recipes for Data Visualization}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of custom ggplot2-based visualizations for data exploration and analysis. Each function handles data preprocessing and returns a ggplot2 object that can be further customized using standard ggplot2 syntax.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ignophi/ggrecipes}
  \item \url{https://ignophi.github.io/ggrecipes/}
  \item Report bugs at \url{https://github.com/Ignophi/ggrecipes/issues}
}

}
\author{
\strong{Maintainer}: Ignophi Hu \email{ignophi.hu@pm.me} (\href{https://orcid.org/0009-0002-4094-0868}{ORCID})

}
\keyword{internal}
